enhance: Memoize buildQueryKey()
