+++ packages/endpoint/src/schemas/Array.ts
@@ -47,8 +47,8 @@ export default class ArraySchema extends PolymorphicSchema {
   queryKey(
     args: unknown,
     queryKey: unknown,
-    lookupIndex: unknown,
     lookupEntities: unknown,
+    lookupIndex: unknown,
   ): any {
     return undefined;
   }
