+++ packages/normalizr/src/interface.ts
@@ -10,8 +10,8 @@ export interface Queryable {
   queryKey(
     args: readonly any[],
     queryKey: (...args: any) => any,
-    lookupIndex: LookupIndex,
     lookupEntities: LookupEntities,
+    lookupIndex: LookupIndex,
     // Must be non-void
   ): {};
 }
@@ -39,8 +39,8 @@ export interface SchemaSimple<T = any, Args extends any[] = any[]> {
   queryKey(
     args: Args,
     queryKey: (...args: any) => any,
-    lookupIndex: LookupIndex,
     lookupEntities: LookupEntities,
+    lookupIndex: LookupIndex,
   ): any;
 }
 
@@ -103,6 +103,9 @@ export interface LookupEntities {
 }
 /** Get PK using an Entity Index */
 export interface LookupIndex {
-  /** lookupIndex('User', 'username', 'ntucker') */
-  (entityKey: string, indexName: string, indexKey: string): string | undefined;
+  /** lookupIndex('User', 'username')['ntucker'] */
+  (
+    entityKey: string,
+    indexName: string,
+  ): { readonly [indexKey: string]: string | undefined };
 }
