+++ website/src/components/Playground/editor-types/@data-client/rest.d.ts
@@ -65,7 +65,7 @@ interface EntityMap<T = any> {
     readonly [k: string]: EntityInterface<T>;
 }
 type SchemaArgs<S extends Schema> = S extends EntityInterface<infer U> ? [EntityFields<U>] : S extends ({
-    queryKey(args: infer Args, indexes: any, recurse: (...args: any) => any, entities: any): any;
+    queryKey(args: infer Args, queryKey: (...args: any) => any, lookupEntities: any, lookupIndex: any): any;
 }) ? Args : never;
 
 interface SnapshotInterface {
@@ -131,7 +131,7 @@ type Schema = null | string | {
     [K: string]: any;
 } | Schema[] | SchemaSimple | Serializable;
 interface Queryable {
-    queryKey(args: readonly any[], indexes: NormalizedIndex, recurse: (...args: any) => any, entities: EntityTable): {};
+    queryKey(args: readonly any[], queryKey: (...args: any) => any, lookupEntities: LookupEntities, lookupIndex: LookupIndex): {};
 }
 type Serializable<T extends {
     toJSON(): string;
@@ -141,7 +141,7 @@ type Serializable<T extends {
 interface SchemaSimple<T = any, Args extends readonly any[] = any[]> {
     normalize(input: any, parent: any, key: any, visit: (...args: any) => any, addEntity: (...args: any) => any, visitedEntities: Record<string, any>, storeEntities: any, args: any[]): any;
     denormalize(input: {}, args: readonly any[], unvisit: (input: any, schema: any) => any): T;
-    queryKey(args: Args, indexes: NormalizedIndex, recurse: (...args: any) => any, entities: EntityTable): any;
+    queryKey(args: Args, queryKey: (...args: any) => any, lookupEntities: LookupEntities, lookupIndex: LookupIndex): any;
 }
 interface SchemaClass<T = any, N = T | undefined, Args extends any[] = any[]> extends SchemaSimple<T, Args> {
     _normalizeNullable(): any;
@@ -164,17 +164,18 @@ interface PolymorphicInterface<T = any, Args extends any[] = any[]> extends Sche
     _normalizeNullable(): any;
     _denormalizeNullable(): any;
 }
-interface NormalizedIndex {
-    readonly [entityKey: string]: {
-        readonly [indexName: string]: {
-            readonly [lookup: string]: string;
-        };
-    };
-}
-interface EntityTable {
-    [entityKey: string]: {
-        [pk: string]: unknown;
+/** Get Array of entities with map function applied */
+interface LookupEntities {
+    (entityKey: string): {
+        readonly [pk: string]: any;
     } | undefined;
+}
+/** Get PK using an Entity Index */
+interface LookupIndex {
+    /** lookupIndex('User', 'username', 'ntucker') */
+    (entityKey: string, indexName: string): {
+        readonly [indexKey: string]: string | undefined;
+    };
 }
 /** Defines a networking endpoint */
 interface EndpointInterface<F extends FetchFunction = FetchFunction, S extends Schema | undefined = Schema | undefined, M extends true | undefined = true | undefined> extends EndpointExtraOptions<F> {
@@ -397,7 +398,7 @@ interface IEntityClass<TBase extends Constructor = any> {
      *
      * @see https://dataclient.io/rest/api/Entity#queryKey
      */
-    queryKey(args: readonly any[], indexes: NormalizedIndex, recurse: any, entities: any): any;
+    queryKey(args: readonly any[], queryKey: any, lookupEntities: LookupEntities, lookupIndex: LookupIndex): any;
     denormalize<T extends (abstract new (...args: any[]) => IEntityInstance & InstanceType<TBase>) & IEntityClass & TBase>(this: T, input: any, args: readonly any[], unvisit: (input: any, schema: any) => any): AbstractInstanceType<T>;
     /** All instance defaults set */
     readonly defaults: any;
@@ -444,7 +445,7 @@ declare class Invalidate<E extends EntityInterface & {
         fetchedAt: number;
     };
     /** /End Normalize lifecycles **/
-    queryKey(args: any, indexes: any, recurse: any): undefined;
+    queryKey(args: any, queryKey: unknown, lookupEntities: unknown, lookupIndex: unknown): undefined;
     denormalize(id: string, args: readonly any[], unvisit: (input: any, schema: any) => any): AbstractInstanceType<E>;
     _denormalizeNullable(): AbstractInstanceType<E> | undefined;
     _normalizeNullable(): string | undefined;
@@ -461,7 +462,7 @@ declare class Query<S extends Queryable, P extends (entries: DenormalizeNullable
     constructor(schema: S, process: P);
     normalize(...args: any): any;
     denormalize(input: {}, args: any, unvisit: any): ReturnType<P> | undefined;
-    queryKey(args: ProcessParameters<P, S>, indexes: any, recurse: (schema: any, args: any, indexes: NormalizedIndex, entities: EntityTable) => any, entities: EntityTable): any;
+    queryKey(args: ProcessParameters<P, S>, queryKey: (schema: any, args: any, lookupEntities: LookupEntities, lookupIndex: LookupIndex) => any, lookupEntities: LookupEntities, lookupIndex: LookupIndex): any;
     _denormalizeNullable: (input: {}, args: readonly any[], unvisit: (input: any, schema: any) => any) => ReturnType<P> | undefined;
     _normalizeNullable: () => NormalizeNullable<S>;
 }
@@ -516,7 +517,7 @@ interface CollectionInterface<S extends PolymorphicInterface = any, Args extends
         date: number;
         fetchedAt: number;
     };
-    queryKey(args: Args, indexes: unknown, recurse: unknown, entities: unknown): any;
+    queryKey(args: Args, queryKey: unknown, lookupEntities: unknown, lookupIndex: unknown): any;
     createIfValid: (value: any) => any | undefined;
     denormalize(input: any, args: readonly any[], unvisit: (input: any, schema: any) => any): ReturnType<S['denormalize']>;
     _denormalizeNullable(): ReturnType<S['_denormalizeNullable']>;
@@ -592,9 +593,9 @@ declare class Array$1<S extends Schema = Schema> implements SchemaClass {
 
   queryKey(
     args: readonly any[],
-    indexes: NormalizedIndex,
-    recurse: (...args: any) => any,
-    entities: any,
+    queryKey: (...args: any) => any,
+    lookupEntities: any,
+    lookupIndex: any,
   ): undefined;
 }
 
@@ -643,9 +644,9 @@ declare class All<
   queryKey(
     // TODO: hack for now to allow for variable arg combinations with Query
     args: [] | [unknown],
-    indexes: NormalizedIndex,
-    recurse: (...args: any) => any,
-    entities: EntityTable,
+    queryKey: (...args: any) => any,
+    lookupEntities: LookupEntities,
+    lookupIndex: LookupIndex,
   ): any;
 }
 
@@ -682,9 +683,9 @@ declare class Object$1<O extends Record<string, any> = Record<string, Schema>>
 
   queryKey(
     args: readonly any[],
-    indexes: NormalizedIndex,
-    recurse: (...args: any) => any,
-    entities: any,
+    queryKey: (...args: any) => any,
+    lookupEntities: LookupEntities,
+    lookupIndex: LookupIndex,
   ): any;
 }
 
@@ -766,9 +767,9 @@ interface UnionInstance<
 
   queryKey(
     args: [Args],
-    indexes: NormalizedIndex,
-    recurse: (...args: any) => any,
-    entities: any,
+    queryKey: (...args: any) => any,
+    lookupEntities: LookupEntities,
+    lookupIndex: LookupIndex,
   ): { id: any; schema: string };
 }
 
@@ -850,9 +851,9 @@ declare class Values<Choices extends Schema = any> implements SchemaClass {
 
   queryKey(
     args: readonly any[],
-    indexes: NormalizedIndex,
-    recurse: (...args: any) => any,
-    entities: any,
+    queryKey: (...args: any) => any,
+    lookupEntities: LookupEntities,
+    lookupIndex: LookupIndex,
   ): undefined;
 }
